pivot to side quests
