+++ src/constants/sideQuestOptions.ts
@@ -1,20 +1,20 @@
 import {
-  PrivateChallengeAvoidType,
-  PrivateChallengeGoal,
-  PrivateChallengeHardSituation,
-  PrivateChallengeMeaningfulType,
-  PrivateChallengePreferredContext,
-  PrivateChallengeProgressDefinition,
-  PrivateChallengeQuestionnaireDraft,
-  PrivateChallengeStretchLevel,
-} from "../types/privateChallenges";
+  SideQuestAvoidType,
+  SideQuestGoal,
+  SideQuestHardSituation,
+  SideQuestMeaningfulType,
+  SideQuestPreferredContext,
+  SideQuestProgressDefinition,
+  SideQuestQuestionnaireDraft,
+  SideQuestStretchLevel,
+} from "../types/sideQuests";
 
 type Option<T extends string> = {
   id: T;
   label: string;
 };
 
-export const PRIVATE_CHALLENGE_EMPTY_DRAFT: PrivateChallengeQuestionnaireDraft = {
+export const SIDE_QUEST_EMPTY_DRAFT: SideQuestQuestionnaireDraft = {
   goal: [],
   hard_situation: [],
   stretch_level: null,
@@ -24,7 +24,7 @@ export const PRIVATE_CHALLENGE_EMPTY_DRAFT: PrivateChallengeQuestionnaireDraft =
   progress_definition: [],
 };
 
-export const PRIVATE_CHALLENGE_GOAL_OPTIONS: Option<PrivateChallengeGoal>[] = [
+export const SIDE_QUEST_GOAL_OPTIONS: Option<SideQuestGoal>[] = [
   { id: "novelty", label: "Novelty" },
   { id: "fun", label: "Fun" },
   { id: "connection", label: "Connection" },
@@ -33,7 +33,7 @@ export const PRIVATE_CHALLENGE_GOAL_OPTIONS: Option<PrivateChallengeGoal>[] = [
   { id: "better_stories", label: "Better stories" },
 ];
 
-export const PRIVATE_CHALLENGE_HARD_SITUATION_OPTIONS: Option<PrivateChallengeHardSituation>[] = [
+export const SIDE_QUEST_HARD_SITUATION_OPTIONS: Option<SideQuestHardSituation>[] = [
   { id: "low_energy", label: "Low energy" },
   { id: "overthinking", label: "Overthinking" },
   { id: "spending_money", label: "Not wanting to spend money" },
@@ -44,13 +44,13 @@ export const PRIVATE_CHALLENGE_HARD_SITUATION_OPTIONS: Option<PrivateChallengeHa
   { id: "feeling_self_conscious", label: "Feeling self-conscious" },
 ];
 
-export const PRIVATE_CHALLENGE_STRETCH_LEVEL_OPTIONS: Option<PrivateChallengeStretchLevel>[] = [
+export const SIDE_QUEST_STRETCH_LEVEL_OPTIONS: Option<SideQuestStretchLevel>[] = [
   { id: "easy_win", label: "Easy win" },
   { id: "moderate_push", label: "Moderate push" },
   { id: "bold_nudge", label: "Bold nudge" },
 ];
 
-export const PRIVATE_CHALLENGE_CONTEXT_OPTIONS: Option<PrivateChallengePreferredContext>[] = [
+export const SIDE_QUEST_CONTEXT_OPTIONS: Option<SideQuestPreferredContext>[] = [
   { id: "at_home", label: "At home" },
   { id: "near_home", label: "Near home" },
   { id: "out_in_the_city", label: "Out in the city" },
@@ -58,7 +58,7 @@ export const PRIVATE_CHALLENGE_CONTEXT_OPTIONS: Option<PrivateChallengePreferred
   { id: "solo", label: "Solo" },
 ];
 
-export const PRIVATE_CHALLENGE_MEANINGFUL_TYPE_OPTIONS: Option<PrivateChallengeMeaningfulType>[] = [
+export const SIDE_QUEST_MEANINGFUL_TYPE_OPTIONS: Option<SideQuestMeaningfulType>[] = [
   { id: "playful", label: "Playful" },
   { id: "creative", label: "Creative" },
   { id: "exploratory", label: "Exploratory" },
@@ -67,7 +67,7 @@ export const PRIVATE_CHALLENGE_MEANINGFUL_TYPE_OPTIONS: Option<PrivateChallengeM
   { id: "growth_edge", label: "A little uncomfortable in a good way" },
 ];
 
-export const PRIVATE_CHALLENGE_AVOID_OPTIONS: Option<PrivateChallengeAvoidType>[] = [
+export const SIDE_QUEST_AVOID_OPTIONS: Option<SideQuestAvoidType>[] = [
   { id: "spending_money", label: "Spending money" },
   { id: "talking_to_strangers", label: "Talking to strangers" },
   { id: "group_social_situations", label: "Group/social situations" },
@@ -77,7 +77,7 @@ export const PRIVATE_CHALLENGE_AVOID_OPTIONS: Option<PrivateChallengeAvoidType>[
   { id: "going_far", label: "Going far" },
 ];
 
-export const PRIVATE_CHALLENGE_PROGRESS_OPTIONS: Option<PrivateChallengeProgressDefinition>[] = [
+export const SIDE_QUEST_PROGRESS_OPTIONS: Option<SideQuestProgressDefinition>[] = [
   { id: "did_something_unusual", label: "I did things I normally would not do" },
   { id: "more_stories", label: "I have more stories and memories" },
   { id: "days_less_repetitive", label: "My days felt less repetitive" },
