+++ src/constants/translations.ts
@@ -399,6 +399,12 @@ export const translations = {
   "Physically demanding activities": "Attività fisicamente impegnative",
   "Nighttime activities": "Attività notturne",
   "I have more stories and memories": "Ho più storie e ricordi",
+  "Side quests for right now": "Side quest per questo momento",
+  "Ranked to fit the kind of break from routine you want right now.": "Ordinate per adattarsi al tipo di deviazione dalla routine che vuoi in questo momento.",
+  "No side quests yet.": "Nessuna side quest ancora.",
+  "Once side quests are added, they will show up here ranked for your preferences.": "Quando aggiungeremo le side quest, appariranno qui ordinate in base alle tue preferenze.",
+  "Why this works": "Perché funziona",
+  "Try it like this": "Provala così",
   "Save preferences": "Salva preferenze",
   "Edit preferences": "Modifica preferenze",
   "Today on your path": "Oggi sul tuo percorso",
