+++ src/services/sideQuestService.ts
@@ -1,30 +1,19 @@
 import { supabase } from "../lib/supabase";
 import {
-  PrivateChallengeAvoidType,
-  PrivateChallenge,
-  PrivateChallengeDifficulty,
-  PrivateChallengeGoal,
-  PrivateChallengeHardSituation,
-  PrivateChallengeMeaningfulType,
-  PrivateChallengePreferredContext,
-  PrivateChallengeProfile,
-  PrivateChallengeProgressDefinition,
-  PrivateChallengeQuestionnaireDraft,
-  PrivateChallengeSet,
-  PrivateChallengeStretchLevel,
-  PrivateChallengeStats,
-} from "../types/privateChallenges";
-
-type ChallengeRow = Omit<PrivateChallenge, "llm_metadata"> & {
-  llm_metadata: Record<string, unknown> | null;
-};
-
-type SetRow = Omit<PrivateChallengeSet, "private_challenges"> & {
-  private_challenges: ChallengeRow[] | null;
-};
+  SideQuest,
+  SideQuestAvoidType,
+  SideQuestGoal,
+  SideQuestHardSituation,
+  SideQuestMeaningfulType,
+  SideQuestPreferredContext,
+  SideQuestProfile,
+  SideQuestProgressDefinition,
+  SideQuestQuestionnaireDraft,
+  SideQuestStretchLevel,
+} from "../types/sideQuests";
 
 type ProfileRow = Omit<
-  PrivateChallengeProfile,
+  SideQuestProfile,
   "goal" | "hard_situation" | "preferred_context" | "meaningful_type" | "progress_definition"
 > & {
   goal: string[] | string | null;
@@ -36,12 +25,24 @@ type ProfileRow = Omit<
   avoid_types: string[] | null;
 };
 
+type SideQuestRow = Omit<
+  SideQuest,
+  "goal_tags" | "barrier_tags" | "context_tags" | "type_tags" | "outcome_tags" | "avoid_flags"
+> & {
+  goal_tags: string[] | null;
+  barrier_tags: string[] | null;
+  context_tags: string[] | null;
+  type_tags: string[] | null;
+  outcome_tags: string[] | null;
+  avoid_flags: string[] | null;
+};
+
 function toArray<T extends string>(value: T[] | T | null | undefined): T[] {
   if (Array.isArray(value)) return value;
   return value ? [value] : [];
 }
 
-const GOAL_MAP: Record<string, PrivateChallengeGoal> = {
+const GOAL_MAP: Record<string, SideQuestGoal> = {
   novelty: "novelty",
   fun: "fun",
   connection: "connection",
@@ -57,7 +58,7 @@ const GOAL_MAP: Record<string, PrivateChallengeGoal> = {
   less_overthinking: "novelty",
 };
 
-const HARD_SITUATION_MAP: Record<string, PrivateChallengeHardSituation> = {
+const HARD_SITUATION_MAP: Record<string, SideQuestHardSituation> = {
   low_energy: "low_energy",
   overthinking: "overthinking",
   spending_money: "spending_money",
@@ -74,7 +75,7 @@ const HARD_SITUATION_MAP: Record<string, PrivateChallengeHardSituation> = {
   saying_yes: "overthinking",
 };
 
-const STRETCH_LEVEL_MAP: Record<string, PrivateChallengeStretchLevel> = {
+const STRETCH_LEVEL_MAP: Record<string, SideQuestStretchLevel> = {
   easy_win: "easy_win",
   moderate_push: "moderate_push",
   bold_nudge: "bold_nudge",
@@ -83,7 +84,7 @@ const STRETCH_LEVEL_MAP: Record<string, PrivateChallengeStretchLevel> = {
   push_me: "bold_nudge",
 };
 
-const CONTEXT_MAP: Record<string, PrivateChallengePreferredContext | PrivateChallengePreferredContext[]> = {
+const CONTEXT_MAP: Record<string, SideQuestPreferredContext | SideQuestPreferredContext[]> = {
   at_home: "at_home",
   near_home: "near_home",
   out_in_the_city: "out_in_the_city",
@@ -95,7 +96,7 @@ const CONTEXT_MAP: Record<string, PrivateChallengePreferredContext | PrivateChal
   work_or_school: "near_home",
 };
 
-const MEANINGFUL_TYPE_MAP: Record<string, PrivateChallengeMeaningfulType> = {
+const MEANINGFUL_TYPE_MAP: Record<string, SideQuestMeaningfulType> = {
   playful: "playful",
   creative: "creative",
   exploratory: "exploratory",
@@ -110,7 +111,7 @@ const MEANINGFUL_TYPE_MAP: Record<string, PrivateChallengeMeaningfulType> = {
   habit_building: "reflective",
 };
 
-const AVOID_TYPE_MAP: Record<string, PrivateChallengeAvoidType> = {
+const AVOID_TYPE_MAP: Record<string, SideQuestAvoidType> = {
   spending_money: "spending_money",
   talking_to_strangers: "talking_to_strangers",
   group_social_situations: "group_social_situations",
@@ -121,7 +122,7 @@ const AVOID_TYPE_MAP: Record<string, PrivateChallengeAvoidType> = {
   work_or_school: "lots_of_planning",
 };
 
-const PROGRESS_DEFINITION_MAP: Record<string, PrivateChallengeProgressDefinition> = {
+const PROGRESS_DEFINITION_MAP: Record<string, SideQuestProgressDefinition> = {
   did_something_unusual: "did_something_unusual",
   more_stories: "more_stories",
   days_less_repetitive: "days_less_repetitive",
@@ -166,7 +167,7 @@ function mapValuesExpanded<T extends string>(
   }, []);
 }
 
-function normalizeProfile(row: ProfileRow): PrivateChallengeProfile {
+function normalizeProfile(row: ProfileRow): SideQuestProfile {
   return {
     ...row,
     goal: mapValues(row.goal, GOAL_MAP),
@@ -179,89 +180,24 @@ function normalizeProfile(row: ProfileRow): PrivateChallengeProfile {
   };
 }
 
-function normalizeSet(row: SetRow): PrivateChallengeSet {
-  const order: Record<PrivateChallengeDifficulty, number> = {
-    easy: 0,
-    medium: 1,
-    hard: 2,
-  };
-
+function normalizeSideQuest(row: SideQuestRow): SideQuest {
   return {
     ...row,
-    private_challenges: (row.private_challenges || [])
-      .map((challenge) => ({
-        ...challenge,
-        llm_metadata: challenge.llm_metadata || {},
-      }))
-      .sort((a, b) => order[a.difficulty] - order[b.difficulty]),
+    goal_tags: mapValues(row.goal_tags, GOAL_MAP),
+    barrier_tags: mapValues(row.barrier_tags, HARD_SITUATION_MAP),
+    context_tags: mapValuesExpanded(row.context_tags, CONTEXT_MAP),
+    type_tags: mapValues(row.type_tags, MEANINGFUL_TYPE_MAP),
+    outcome_tags: mapValues(row.outcome_tags, PROGRESS_DEFINITION_MAP),
+    avoid_flags: mapValues(row.avoid_flags, AVOID_TYPE_MAP),
+    stretch_level: STRETCH_LEVEL_MAP[row.stretch_level] || "moderate_push",
+    night_level: row.night_level === 1 ? 1 : 0,
   };
 }
 
-function toDateParts(date: string) {
-  const [year, month, day] = date.split("-").map(Number);
-  return { year, month, day };
-}
-
-function diffDays(a: string, b: string) {
-  const aParts = toDateParts(a);
-  const bParts = toDateParts(b);
-  const aUtc = Date.UTC(aParts.year, aParts.month - 1, aParts.day);
-  const bUtc = Date.UTC(bParts.year, bParts.month - 1, bParts.day);
-  return Math.round((aUtc - bUtc) / (1000 * 60 * 60 * 24));
-}
-
-export function getLocalDateString(date = new Date()) {
-  const year = date.getFullYear();
-  const month = `${date.getMonth() + 1}`.padStart(2, "0");
-  const day = `${date.getDate()}`.padStart(2, "0");
-  return `${year}-${month}-${day}`;
-}
-
-export function computePrivateChallengeStats(sets: PrivateChallengeSet[]): PrivateChallengeStats {
-  const completedDates = sets
-    .filter((set) => set.status === "completed")
-    .map((set) => set.challenge_date)
-    .sort((a, b) => b.localeCompare(a));
-
-  if (completedDates.length === 0) {
-    return {
-      currentStreak: 0,
-      longestStreak: 0,
-      totalCompleted: 0,
-    };
-  }
-
-  let currentStreak = 1;
-  for (let i = 1; i < completedDates.length; i++) {
-    if (diffDays(completedDates[i - 1], completedDates[i]) === 1) {
-      currentStreak += 1;
-    } else {
-      break;
-    }
-  }
-
-  let longestStreak = 1;
-  let running = 1;
-  for (let i = 1; i < completedDates.length; i++) {
-    if (diffDays(completedDates[i - 1], completedDates[i]) === 1) {
-      running += 1;
-      longestStreak = Math.max(longestStreak, running);
-    } else {
-      running = 1;
-    }
-  }
-
-  return {
-    currentStreak,
-    longestStreak,
-    totalCompleted: completedDates.length,
-  };
-}
-
-export const privateChallengeService = {
-  async fetchProfile(userId: string): Promise<PrivateChallengeProfile | null> {
+export const sideQuestService = {
+  async fetchProfile(userId: string): Promise<SideQuestProfile | null> {
     const { data, error } = await supabase
-      .from("private_challenge_profiles")
+      .from("side_quest_profiles")
       .select("*")
       .eq("user_id", userId)
       .maybeSingle();
@@ -270,7 +206,7 @@ export const privateChallengeService = {
     return data ? normalizeProfile(data as ProfileRow) : null;
   },
 
-  async saveProfile(userId: string, draft: PrivateChallengeQuestionnaireDraft): Promise<PrivateChallengeProfile> {
+  async saveProfile(userId: string, draft: SideQuestQuestionnaireDraft): Promise<SideQuestProfile> {
     if (
       draft.goal.length === 0 ||
       draft.hard_situation.length === 0 ||
@@ -286,7 +222,7 @@ export const privateChallengeService = {
       user_id: userId,
       goal: draft.goal,
       hard_situation: draft.hard_situation,
-      stretch_level: draft.stretch_level as PrivateChallengeStretchLevel,
+      stretch_level: draft.stretch_level as SideQuestStretchLevel,
       preferred_context: draft.preferred_context,
       meaningful_type: draft.meaningful_type,
       avoid_types: draft.avoid_types,
@@ -296,7 +232,7 @@ export const privateChallengeService = {
     };
 
     const { data, error } = await supabase
-      .from("private_challenge_profiles")
+      .from("side_quest_profiles")
       .upsert(payload, { onConflict: "user_id" })
       .select("*")
       .single();
@@ -305,66 +241,14 @@ export const privateChallengeService = {
     return normalizeProfile(data as ProfileRow);
   },
 
-  async fetchOrCreateTodaySet(localDate: string, language: "en" | "it"): Promise<PrivateChallengeSet> {
-    const { data, error } = await supabase.functions.invoke("get-private-challenge-set", {
-      body: {
-        localDate,
-        language,
-      },
-    });
-
-    if (error) throw error;
-    return normalizeSet(data as SetRow);
-  },
-
-  async fetchHistory(userId: string, limit = 30): Promise<PrivateChallengeSet[]> {
+  async fetchSideQuests(): Promise<SideQuest[]> {
     const { data, error } = await supabase
-      .from("private_challenge_sets")
-      .select("*, private_challenges(*)")
-      .eq("user_id", userId)
-      .order("challenge_date", { ascending: false })
-      .limit(limit);
+      .from("side_quests")
+      .select("*")
+      .eq("is_active", true)
+      .order("id", { ascending: true });
 
     if (error) throw error;
-    return (data || []).map((row) => normalizeSet(row as SetRow));
-  },
-
-  async completeSet(
-    setId: number,
-    difficulty: PrivateChallengeDifficulty,
-    note: string
-  ): Promise<void> {
-    const { data, error } = await supabase
-      .from("private_challenge_sets")
-      .update({
-        status: "completed",
-        completed_difficulty: difficulty,
-        completed_note: note.trim() ? note.trim() : null,
-        completed_at: new Date().toISOString(),
-        updated_at: new Date().toISOString(),
-      })
-      .select("id")
-      .eq("id", setId)
-      .eq("status", "pending")
-      .maybeSingle();
-
-    if (error) throw error;
-    if (!data) throw new Error("This challenge is no longer available");
-  },
-
-  async skipSet(setId: number): Promise<void> {
-    const { data, error } = await supabase
-      .from("private_challenge_sets")
-      .update({
-        status: "skipped",
-        updated_at: new Date().toISOString(),
-      })
-      .select("id")
-      .eq("id", setId)
-      .eq("status", "pending")
-      .maybeSingle();
-
-    if (error) throw error;
-    if (!data) throw new Error("This challenge is no longer available");
+    return (data || []).map((row) => normalizeSideQuest(row as SideQuestRow));
   },
 };
